LayoutTests/http/tests/cookies/partitioned/set-partitioned-and-non-partitioned-cookie-expected.txt

Tests that setting Partitioned and non-partitioned cookie from non-secure document are rejected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


cookie-js-not-secure.html: Found 0 cookies:
cookie-js-secure.html: Found 0 cookies:
cookie-http-header-not-secure.html: Found 0 cookies:
cookie-http-header-secure.html: Found 0 cookies:
httponly-cookie-http-header-not-secure.html: Found 0 cookies:
httponly-cookie-http-header-secure.html: Found 0 cookies:
httponly-cookie-js-not-secure.html: Found 0 cookies:
httponly-cookie-js-secure.html: Found 0 cookies:
PASS successfullyParsed is true

TEST COMPLETE
